utils: Move archive() from Utils to ArchiveUtils
`ArchiveUtils` is  a better place for the function as it contains
related helper functions.
